Fixed 6279666 ("[devtest] Regression: Unit test for 4655010 is failling on PE/EE").
Regression had been caused by fix for 6254469 ("Japanese character is
corrupted when displaying error page", ported from iAS7), which
hard-codes the encoding of error/exception messages to UTF-8. This is
usually good enough, but unit test was stricter and was expecting
the response encoding ("Shift_JIS") that had been set in the test servlet.

Fixed both issues by always using the response encoding (set by the
servlet/JSP) for any error/exception messages. This already used to
work for 500 responses caused by call to response.sendError(), but
wasn't working for 500 responses caused by an exception. In the latter
case, the response is reset, causing its encoding to be reinitialized
to the default (ISO-8859-1). Fixed this by saving the response encoding
before call to reset(), and restoring it afterwards.

Reviewed by: Jeanfrancois Arcand

Tests run:
- JSP TCK
- appserv-tests/devtests/web/responseErrorMessageEncoding (FIXED)
- appserv-tests/devtests/web/responseExceptionMessageEncoding (NEW)


Original author: jluehe
Date: 2005-07-26 01:56:30+00:00