Restored original stack depth from WS 6.x by replacing calls to
  wrapper.invoke(request, response);
with
  wrapper.getPipeline().invoke(request, response);

and

  context.invoke(request, response);
with
  context.getPipeline().invoke(request, response);

in the respective standard valves.

The check for ad-hoc paths is now performed in a new WebPipeline
(subclass of StandardPipeline), which is associated with web modules
in AS and returned by their getPipeline().

Web modules in WS will continue to have Tomcat's StandardPipeline
superclass associated with them.

Tests run:
- Servlet TCK
- QL/PE


Original author: jluehe
Date: 2005-08-18 02:30:13+00:00