Workaround for JDK bug 6448723: SSLContext + X509KeyManager implementation cause SSLHandshakeException: no cipher suites in common
Implementing the X509KeyManager only works when SSL is used with blocking socket. When used with non blocking socket, the JSSE implementation isn't able to invoke chooseEngineServerAlias and chooseEngineClientAlias with the SSLEngine (The API is missing on X509KeyManager), leaving the SSLEngine in an inconsistent state, later throwing:

Caused by: javax.net.ssl.SSLHandshakeException: no cipher suites in common
        at com.sun.net.ssl.internal.ssl.Alerts.getSSLException(Alerts.java:150)
        at com.sun.net.ssl.internal.ssl.SSLEngineImpl.fatal(SSLEngineImpl.java:1352)
        at com.sun.net.ssl.internal.ssl.Handshaker.fatalSE(Handshaker.java:176)
        at com.sun.net.ssl.internal.ssl.Handshaker.fatalSE(Handshaker.java:164)

which is not right. To workaround the problem, I've extended X509ExtendedKeyManager instead and boom everything works!

- ALL QL passed
- Tested Admin Gui with SSL.

Reviewed by: Jan Luehe

[NEWS:Grizzly JDK 5.0/6.0 ea-b10] Workaround for JDK bug 6448723


Original author: jfarcand
Date: 2006-07-13 22:43:40+00:00