Related to https://glassfish.dev.java.net/issues/show_bug.cgi?id=836:
Reverted my previous commit. The check for the "isSkippable" condition is already pretty high up in the SSO valve's invoke() logic, so determining whether to skip the valve, and possibly skip it, would not have resulted in any significant performance gains. Also, the code was adding extra overhead when checking if the SSO valve's postInvoke() method needed to be invoked (it is easier to just call this method, which does nothing).

I've optimized the SSO valve's invoke() method instead: For some reason that I don't understand, the code was allocating a cookie[] if there were no cookies present in the request (this code has been around for the longest time). Changed the code to stop allocating cookie[] and to return INVOKE_NEXT immediately.

Reviewed by: Larry White, Jeanfrancois Arcand

Tests run: QL/PE


Original author: jluehe
Date: 2006-07-19 21:40:53+00:00