Implemented suggested fix proposed by Santiago Pericas-Geertsen to avoid 6402288 ("Impossible to create javax.xml.validation.Schema instance from web-app_2_4.xsd").
The fix was to set the schema's systemId on the org.w3c.dom.ls.LSInput
in org.apache.jasper.xmlparser.ParserUtils:

  LSInput.setSystemId(systemId);

so that Xerces is able to detect any loops.

With this, the XML validation flag in Jasper can now be used for both
TLD and web.xml validations, as originally intended.

Tests run:
- QL/PE
- JSP/TCK


Original author: jluehe
Date: 2006-08-01 22:28:49+00:00