Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=862 ("JSR-88 DeploymentManager.redeploy fails for ear"):
Have StandardHost.findChild() decode its context root argument before
attempting to match it against any of the deployed contexts.
This is to match the behaviour of StandardContext.setPath(), which also
decodes it context root argument before setting it as the context's
path (which is also used as the context's name).

Tests run: QL


Original author: jluehe
Date: 2006-08-16 20:04:25+00:00