I keep forgetting to include the NEWS tag. :(
[NEWS:webcontainer 896] OPTIMIZATION: Do not include cookie for invalidated HTTP session in the response
Created a new class: org.apache.coyote.tomcat5.SessionTracker, which is
responsbile for keeping track of the total number of active sessions
associated with an HTTP request.
A given HTTP request may be associated with more than one active
session if it has been dispatched to one or more foreign contexts.
The number of active sessions being tracked is used to determine
whether or not any session information needs to be reflected in the
response (e.g., in the form of a response cookie). If no active
sessions are associated with the request by the time the response is
committed, no session information will be reflected in the response.
A SessionTracker tracks a session by adding itself as an
org.apache.catalina.SessionListener to it (this is the same interface
implemented by our SSO valve). On receiving a SESSION_DESTROYED event,
it decrements its counter of tracked sessions.
Delayed cookie generation to the time when the response is
committed: A cookie is generated and added to the response only if the
number of session's tracked by the request's SessionTacker is > 0.
Each request has an instance of SessionTracker as one of its
instance vars. A SessionTracker is reset when its request is getting
recycled. When a request is put to service, it adds its SessionTracker
instance to its request attributes (under the attribute name
"com.sun.enterprise.http.sessionTracker"),
thereby making it available (via a call to getAttribute())
to any request wrappers that are wrapped around the request during a
RequestDispatcher.dispatch operation.


Original author: jluehe
Date: 2006-08-17 16:19:08+00:00