Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=1505 ("Illegal class loader binding happens unexpectedly")
Reviewed by: jfarcand

Tests run: QL, Servlet/TCK

[NEWS:webcontainer 1505] Illegal class loader binding happens unexpectedly
Avoid premature removal of classloader-to-resource mapping. Leave a
classloader's resource bindings intact for as long as the classloader
is active: Make WebappLoader.start() the only code responsible for adding
the mapping, and WebappLoader.stop() the only code responsible for
removing it. No other code (such as StandardContext.bindThread() or
unbindThread()) should set or remove this kind of mapping.


Original author: jluehe
Date: 2006-11-22 17:11:00+00:00